+++ package.json
@@ -10,6 +10,7 @@
 	"dependencies": {
 		"react": "16.0.0-beta.5",
 		"react-native": "0.49.3",
+		"react-navigation": "^1.5.2",
 		"react-viro": "2.4.0"
 	},
 	"devDependencies": {
